Fixing the Faluire Bug

+++ lib/java.rb
@@ -45,8 +45,8 @@ class Java < Code
       $result = `timeout 4s java Code`
 
       if $result.chomp != tc.output
-        File.delete("Code.java")
-        File.delete("Code.class")
+        File.delete("Code.java") if File.exist?("Code.java")
+        File.delete("Code.class") if File.exist?("Code.class")
         return false
       end
